order added

+++ stock-management/stock-management/src/main/java/com/inaya/stockmanagement/model/OrderDetails.java
@@ -6,6 +6,8 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @Entity
 @NoArgsConstructor
@@ -24,7 +26,7 @@ public class OrderDetails {
     private int quantity;
 
     @Column(name = "unit_price")
-    private int unitPrice;
+    private BigDecimal unitPrice;
 
     @JsonIgnore
     @ManyToOne(fetch = FetchType.LAZY)
